+++ js/ext/angular/test/directive/ionicToggle.unit.js
@@ -9,14 +9,14 @@ describe('Ionic Toggle', function() {
     el = $compile('<toggle ng-model="data.name"></toggle>')($rootScope);
   }));
 
-  it('Should load', function() {
+  xit('Should load', function() {
     var toggleView = el.isolateScope().toggle;
     expect(toggleView).not.toEqual(null);
     expect(toggleView.checkbox).not.toEqual(null);
     expect(toggleView.handle).not.toEqual(null);
   });
 
-  it('Should toggle', function() {
+  xit('Should toggle', function() {
     var toggle = el.isolateScope().toggle;
     expect(toggle.val()).toBe(false);
     el.click();
@@ -25,7 +25,7 @@ describe('Ionic Toggle', function() {
     expect(toggle.val()).toBe(false);
   });
 
-  it('Should disable and enable', function() {
+  xit('Should disable and enable', function() {
 
     rootScope.data = { isDisabled: false };
     el = compile('<toggle ng-model="data.name" ng-disabled="data.isDisabled"></toggle>')(rootScope);
@@ -45,4 +45,5 @@ describe('Ionic Toggle', function() {
     el.click();
     expect(toggle.el.getAttribute('disabled')).not.toBe('disabled');
   });
+
 });